# src/openapi/urlShortener.yml
openapi: 3.0.0

info:
  title: URL Shortener OAS
  description: OpenApi specification for a url shortening service
  version: 1.0.0

servers:
  - url: http://localhost:3000/

components:
  schemas:
    urlRequest:
      type: object
      properties:
        user:
          type: string
          format: uuid
        url:
          type: string
      required:
        - user
        - url


    urlResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          type: integer
          format: int64
        url:
          type: string
        shortURL:
          type: string
        created_at:
          type: integer
          format: int64

      required:
        - id
        - user
        - url
        - shortURL
        - created_at


    ShortUrlRequest:
      type: object
      properties:
        url:
          type: string
      required:
        - url


    ShortUrlResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url

    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
      required:
        - code
        - message


paths:
  /urls:
    post:
      description: Creates a new short url 
      operationId: shortenUrl
      requestBody:
        description: Shorten url
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/urlRequest'
      responses:
        '201': 
          description: Shortened url creation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/urlResponse'
        default:
          description: unexpected error 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /urls/{URL}:
    get:
      description: Gets a shortened url
      operationId: getLongURLFromShort
      parameters:
        - in: path
          name: URL
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Shortened Url Exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortUrlResponse"
        '404':
          description: Url doesn's exist 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              